+++ test/oscope/headless_test_runner.clj
@@ -29,6 +29,7 @@
             [oscope.typed-schema-config-test]
             [oscope.typed-schema-runtime-test]
             [oscope.typed-schema-test]
+            [oscope.typed-span-aggregate-test]
             [oscope.typed-span-filter-test]
             [oscope.visualization-editor-test]
             [oscope.web-test]
@@ -60,6 +61,7 @@
                         'oscope.typed-schema-config-test
                         'oscope.typed-schema-runtime-test
                         'oscope.typed-schema-test
+                        'oscope.typed-span-aggregate-test
                         'oscope.typed-span-filter-test
                         'oscope.visualization-editor-test
                         'oscope.web-test